radio-settings: Do not use enum ofono_access_mode
Originally the enum was only meant to signify a single radio access mode
preference, but over time its meaning was overloaded to be a bitfield
with multiple preferences.  Switch away from using an enum value to
using an unsigned int to make it clearer that the value is a bitfield.